+++ src/routes/(app)/eventos/[slug]/+page.svelte
@@ -117,7 +117,9 @@
     {/if}
 
     <div
-      class={`min-w-[350px] mx-auto mt-20 h-min ${!event?.active ? 'max-w-6xl' : 'max-w-6xl' }`}
+      class={`min-w-[350px] mx-auto mt-20 h-min ${
+        !event?.active ? "max-w-6xl" : "max-w-6xl"
+      }`}
     >
       <div
         id="event"
@@ -196,9 +198,7 @@
                   {/if}
                 </div>
                 {#if event?.assistance && !event?.active}
-                  <AttendanceStat
-                    assistance={event?.assistance}
-                  />
+                  <AttendanceStat assistance={event?.assistance} />
                 {/if}
               {/if}
 
@@ -398,7 +398,7 @@
           <h3
             class="text-4xl text-center font-bold text-white mask font-ibm italic mb-4 pl-4"
           >
-            Galería de imagenes
+            Momentos
           </h3>
         </div>
         <div
@@ -409,9 +409,16 @@
       </section>
     {/if}
   </div>
+  <!-- VIDEO RECAP -->
+  {#if event?.recap_video && !event.active}
+    <Video
+      title={event?.recap_video?.title}
+      url={event?.recap_video?.url}
+    />
+  {/if}
   {#if event.videoUrl && !event.active}
     <section class="my-20">
-      <Video title="Streaming del evento" url={event.videoUrl} />
+      <Video title="Streaming" url={event.videoUrl} />
     </section>
   {/if}
   {#if event.playlist && !event.active}